TASK: Centralize signal registry
