+++ Resources/Private/JavaScript/Guest/Sagas/EditorToolbar/index.js
@@ -1,9 +1,9 @@
 import {takeLatest} from 'redux-saga';
 
 import {actionTypes} from 'Guest/Redux/index';
-import {call} from 'Guest/Containers/EditorToolbar/SignalRegistry/index';
+import {call} from 'Guest/Process/SignalRegistry/index';
 
-export function* watchDispatchSignal(getState) {
+export function* watchDispatchSignal() {
     yield* takeLatest(actionTypes.EditorToolbar.DISPATCH_SIGNAL, function* dispatchEditorSignal(action) {
         const signal = action.payload;
